[REFACTORING] PostApi를 ForumApi에서 분리
