+++ src/main/java/com/example/jwt/service/ForumService.java
@@ -18,12 +18,4 @@ public interface ForumService {
     void forumDelete(ForumDto forumDto) throws Exception;
 
     Forum forumPut(ForumChangeDto forumChangeDto) throws Exception;
-
-    List<String> getForumPostList(String forumName) throws Exception;
-
-    Post postCreate(String forumName,PostDto postDto) throws Exception;
-
-    void postDelete(String forumName,PostDto postDto) throws Exception;
-
-    Post postPut(String forumName, PostChangeDto postChangeDto) throws Exception;
 }
