+++ src/test/java/com/example/jwt/service/ForumServiceTest.java
@@ -8,6 +8,7 @@ import com.example.jwt.dto.PostChangeDto;
 import com.example.jwt.dto.PostDto;
 import com.example.jwt.repository.ForumRepository;
 import com.example.jwt.repository.PostRepository;
+import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,14 +22,12 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
+@Slf4j
 @Transactional
 class ForumServiceTest {
 
     @Autowired
     private ForumRepository forumRepository;
-
-    @Autowired
-    private PostRepository postRepository;
 
     @Autowired
     private ForumService forumService;
@@ -91,62 +90,5 @@ class ForumServiceTest {
 
         //then
         assertEquals(forumChangeDto.getNewForumName(),forum.getForumName());
-    }
-
-    @Test
-    void getForumPostList() throws Exception {
-        //given
-        List<String> list;
-
-        //when
-        list = forumService.getForumPostList("민경모씹떢");
-
-        //then
-        for (String i : list){
-            System.out.println(i);
-        }
-    }
-
-    @Test
-    void postCreate() throws Exception{
-        //given
-        PostDto postDto = PostDto.builder()
-                .postName("민경모는 에밀리아")
-                .content("를 좋아하냐????")
-                .build();
-        //when
-        Post post = forumService.postCreate("민경모모모",postDto);
-
-        //then
-        assertEquals(postDto.getPostName(),post.getPostName());
-    }
-
-    @Test
-    void postDelete() throws Exception{
-        //given
-        PostDto postDto = PostDto.builder()
-                .postName("민경모는 에밀리아")
-                .content("를 좋아하냐????")
-                .build();
-        //when
-        forumService.postDelete("민경모모모",postDto);
-
-        //then
-        assertEquals(null,postRepository.findByPostNameAndForum(postDto.getPostName(),forumRepository.findByForumName("민경모모모")));
-    }
-
-    @Test
-    void postPut() throws Exception {
-        //given
-        PostChangeDto postChangeDto = new PostChangeDto();
-        postChangeDto.setPostName("최윤성은 에밀리아");
-        postChangeDto.setNewPostName("민경모는 에밀리아");
-        postChangeDto.setNewContent("를 좋아하냐????");
-
-        //when
-        Post post = forumService.postPut("민경모모모",postChangeDto);
-
-        //then
-        assertEquals(postChangeDto.getNewPostName(),post.getPostName());
     }
 }
